authentication and authorization

+++ base/views.py
@@ -1,13 +1,64 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 # Create your views here.
 
 
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect("home")
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist.")
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "Username or password does not exists.")
+
+    context = {"page": page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect("home")
+
+
+def registerUser(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error occured during registration.")
+    return render(request, 'base/login_register.html', {"form": form})
+
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
-    print("yooo", q, request.GET)
     rooms = Room.objects.filter(
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
@@ -16,16 +67,32 @@ def home(request):
 
     topics = Topic.objects.all()
     room_count = rooms.count()
-    context = {"rooms": rooms, "topics": topics, "room_count": room_count}
+    room_messages = Message.objects.filter(Q(room__topic__name__icontains=q))
+    context = {"rooms": rooms, "topics": topics,
+               "room_count": room_count, "room_messages": room_messages}
     return render(request, 'base/home.html', context)
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {"room": room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get("body")
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
+    context = {"room": room, "room_messages": room_messages,
+               "participants": participants}
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
 
@@ -39,9 +106,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)  # for pre populating data
+
+    if request.user != room.host:
+        return HttpResponse("You'r not allowed here!")
 
     if request.method == 'POST':
         # to see which instance of room is updated
@@ -54,9 +125,26 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You'r not allowed here!")
     if request.method == 'POST':
         room.delete()
         return redirect("home")
     return render(request, 'base/delete.html', {"obj": room})
+
+
+@login_required(login_url="login")
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse("You'r not allowed here!")
+    if request.method == 'POST':
+        message.delete()
+        current_user_messages = Message.objects.filter(user=request.user)
+        if current_user_messages.count() == 0:
+            message.room.participants.remove(request.user)
+        return redirect("home")
+    return render(request, 'base/delete.html', {"obj": message})
